Fetch, crop, and merge source tiles first
A python script reads the tindex and generates an output LAZ file
containing only the points needed for this tile bbox.

This results in a major reduction of memory use by PDAL over using
readers.tindex.  This is because readers.tindex reads all the points
from every source file that touches the bbox.  Additionally,
readers.tindex forces non-streaming mode so all those points remain
in memory for the whole pipeline.

The python script does this by calling PDAL:

First fetching + cropping files individually according to the tindex.
These operations are streaming, so points outside the bbox don't hurt us.

Next the script calls PDAL to merge of the cropped files.  This is
non-streaming, but only includes points within the bbox
(so large source tiles are no longer an issue).

Then finally since the whole DSM, HAG, DEM process is non-streaming,
but it has to deal with the points within the tile.

Note: if only a DEM is needed this step can be streaming too.

Note: using readers.ept has a similar effect, because EPT tiles are small
and regular.  However this still forces non-streaming mode and requires
building an EPT tileset first.